+++ app/src/main/java/ru/akumakeito/currention/repository/PairCurrencyRepositoryImpl.kt
@@ -31,7 +31,7 @@ class PairCurrencyRepositoryImpl @Inject constructor(
 
         var oldPair = getPairById(pairCurrency.id)
 
-        val newRates = convert(pairCurrency.fromCurrency, pairCurrency.toCurrency, 1)
+        val newRates = convert(pairCurrency.fromCurrency, pairCurrency.toCurrency, 1.0)
 
         if (isFromToCurrenciesSame(pairCurrency, oldPair)) {
             oldPair = updateCurrencyPairWithNewCurrencyFromOrTo(pairCurrency)
@@ -75,7 +75,7 @@ class PairCurrencyRepositoryImpl @Inject constructor(
     override suspend fun convert(
         currencyFrom: FiatCurrency,
         currencyTo: FiatCurrency,
-        amount: Int
+        amount: Double
     ) = apiService.getPairRates(
         currencyFrom.shortCode,
         currencyTo.shortCode,
